# backend/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is a sample express server.",
    "version": "1.0.0",
    "title": "Swagger Doc",
    "contact": {
      "email": ""
    },
    "license": {
      "name": "ISC"
    }
  },
  "host": "localhost:3000",
...
  "schemes": [
    "https",
    "http"
  ],
  "paths": {
...
    "/users/{userId}": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Find User by ID",
        "description": "Returns a single user",
        "operationId": "getUserById",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/user"
            }
          },
          "404": {
            "description": "User not found"
          }
        }
      },
      "put": {
        "tags": [
          "users"
        ],
        "summary": "Update user by ID",
        "description": "return update user",
        "operationId": "updateUser",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "User object that needs to be updated",
            "required": true,
            "schema": {
              "$ref": "#/definitions/user"
            }
          },
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "201": {
            "description": "sucsessful",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/user"
              }
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Pet not found"
          }
        }
      }
    },
...
    "/users/me": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Get some information about user",
        "description": "",
        "operationId": "getProfile",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "profile",
            "schema": {
              "$ref": "#/definitions/user"
            }
          },
          "400": {
            "description": "failed operation"
          }
        }
      },
      "delete": {
        "tags": [
          "users"
        ],
        "summary": "Delete user",
        "description": "This can only be done by the logged in user.",
        "operationId": "deleteUser",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "successful"
          },
          "400": {
            "description": "Invalid username supplied"
          },
          "404": {
            "description": "User not found"
          }
        }
      }
    },
...
  },
  "definitions": {
    "user": {
      "type": "object",
      "properties": {
        "login": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "surname": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "role": {
          "type": "string"
        }
      }
    },
...
  }
}